Moved components Ticket and Ticket list. Able to display assigned employees to tickets.
