--- src/app.css
@@ -1,47 +1,0 @@
-.tickets {
-    display: flex;
-    flex-direction: column;
-  }
-
-  .tickets-container {
-    margin: 1rem 5rem 0rem 5rem;
-  }
-
-  .tickets-container > h2 {
-    margin-top: 0;
-  }
-
-  .ticket {
-    background-color: var(--white);
-    border: 1px solid var(--outline);
-    margin: 0.5rem 0;
-    border-radius: 0.5rem;
-    padding: 0.5rem;
-  }
-
-  /* This targets all the child elements of the element with the .ticket class */
-  .ticket > * {
-    padding: 0.5rem;
-  }
-
-  .ticket-info {
-    color: var(--outline);
-    font-family: "Roboto", sans-serif;
-  }
-
-  /* This targets any footer element that is a child of an element with the .ticket class */
-  .ticket > footer {
-    display: flex;
-    border-top: 1px solid var(--outline);
-    align-items: center;
-    justify-content: space-between;
-  }
-
-  /* This targets any element that is a child of a footer that is a child of an element with the .ticket class */
-  .ticket > footer > * {
-    flex-basis: 22%;
-  }
-
-  .filter-btn {
-    margin-right: 2rem;
-  }
