+++ src/services/ticketService.jsx
@@ -1,3 +1,3 @@
 export const getAllTickets = () => {
-    return fetch("http://localhost:8088/serviceTickets").then((res => res.json()))
+    return fetch("http://localhost:8088/serviceTickets?_embed=employeeTickets").then((res => res.json()))
 }
